Добавлен метод, который по пустой области экрана прячет клавиатуру на контролере добавления канала.

+++ iRss/SelectionTableViewController.h
@@ -6,7 +6,7 @@
 @interface SelectionTableViewController : UITableViewController <UITableViewDataSource, UITableViewDelegate> {
     @private
     NSDictionary *dictionaryRssChanel;
-    NSInteger valueKey;
+    int valueKey;
 }
 
 #pragma mark - Action
